+++ src/opencl/activation/relu.cl
@@ -1,11 +1,11 @@
-float activation(float input) {
+double activation(double input) {
     if (input < 0.0) {
         return 0.0;
     }
     return input;
 }
 
-float activationDerivative(float input) {
+double activationDerivative(double input) {
     if (input <= 0.0) {
         return 0.0;
     }
